+++ priv/repo/migrations/20260702121000_remove_import_source_from_gardens.exs
@@ -1,0 +1,9 @@
+defmodule Permaplanner.Repo.Migrations.RemoveImportSourceFromGardens do
+  use Ecto.Migration
+
+  def change do
+    alter table(:gardens) do
+      remove :import_source
+    end
+  end
+end